APP-837 Remove Symphony dependencies from sdk
- Fix PR comments
