+++ integration-provisioning/src/main/java/org/symphonyoss/integration/provisioning/service/ApplicationService.java
@@ -16,7 +16,6 @@
 
 package org.symphonyoss.integration.provisioning.service;
 
-import static org.symphonyoss.integration.provisioning.properties.ApplicationProperties.APP_ID;
 import static org.symphonyoss.integration.provisioning.properties.AuthenticationProperties.DEFAULT_USER_ID;
 
 import com.symphony.api.pod.api.AppEntitlementApi;
@@ -25,7 +24,6 @@ import com.symphony.api.pod.model.PodAppEntitlement;
 import com.symphony.api.pod.model.PodAppEntitlementList;
 import com.symphony.api.pod.model.UserV2;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,8 +32,8 @@ import org.symphonyoss.integration.authentication.AuthenticationProxy;
 import org.symphonyoss.integration.model.yaml.Application;
 import org.symphonyoss.integration.model.yaml.IntegrationProperties;
 import org.symphonyoss.integration.provisioning.client.AppRepositoryClient;
-import org.symphonyoss.integration.provisioning.client.AppStoreBuilder;
-import org.symphonyoss.integration.provisioning.client.AppStoreWrapper;
+import org.symphonyoss.integration.provisioning.client.model.AppStoreBuilder;
+import org.symphonyoss.integration.provisioning.client.model.AppStoreWrapper;
 import org.symphonyoss.integration.provisioning.client.PodApiClientDecorator;
 import org.symphonyoss.integration.provisioning.exception.AppRepositoryClientException;
 import org.symphonyoss.integration.provisioning.exception.ApplicationProvisioningException;
@@ -54,6 +52,8 @@ import javax.annotation.PostConstruct;
 public class ApplicationService {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ApplicationService.class);
+
+  private static final String APP_ID = "id";
 
   @Autowired
   private UserService userService;
